ODPM-419 explained sql_expression.go using comments
